Fix type mismatch in vfs.rs by using AtomicIsize consistently
Changed `io_errno` field from `AtomicI32` to `AtomicIsize` for type
consistency with the `get_errno()` return type.

Changes:
- Import `AtomicIsize` instead of `AtomicI32`
- Update field declaration: `io_errno: AtomicIsize`
- Update initialization: `AtomicIsize::new(0)`
- Remove unnecessary cast in `get_errno()` return

Also updated test assertions to use `isize` instead of `i32` to match
the new return type.

This ensures type consistency throughout the VFS implementation and
eliminates the need for casting between i32 and isize.

+++ ailets-rs/actor_runtime_mocked/tests/vfs_test.rs
@@ -101,7 +101,7 @@ fn read_returns_all_content() {
     let mut buffer = [0u8; 32];
     let bytes_read = vfs.aread(fd, &mut buffer);
 
-    assert_eq!(bytes_read, content.len() as i32);
+    assert_eq!(bytes_read, content.len() as isize);
     assert_eq!(&buffer[..content.len()], content);
 
     // Verify EOF (should return 0 bytes)
@@ -166,7 +166,7 @@ fn write_returns_bytes_written() {
     let content = b"Hello world!";
     let bytes_written = vfs.awrite(fd, content);
 
-    assert_eq!(bytes_written, content.len() as i32);
+    assert_eq!(bytes_written, content.len() as isize);
 
     // Verify written content
     let written = vfs.get_file("test").unwrap();
@@ -184,7 +184,7 @@ fn write_all_content() {
     // Write some content
     let content = b"Hello world!";
     let bytes_written = vfs.awrite(fd, content);
-    assert_eq!(bytes_written, content.len() as i32);
+    assert_eq!(bytes_written, content.len() as isize);
 
     // Verify written content
     let written = vfs.get_file("test").unwrap();
